Run Tests for Customer Dao

+++ src/main/java/com/springapp/customer/CustomerController.java
@@ -19,7 +19,7 @@ public class CustomerController {
     }
 
     @GetMapping("{customerId}")
-    public Customer getCustomerById(@PathVariable("customerId") Integer customerId){
+    public Customer getCustomerById(@PathVariable("customerId") Long customerId){
 
         return customerService.getCustomerById(customerId);
     }
@@ -28,12 +28,12 @@ public class CustomerController {
         customerService.addCustomer(request);
     }
     @DeleteMapping("{customerId}")
-    public void deleteCustomer(@PathVariable("customerId") Integer customerId){
+    public void deleteCustomer(@PathVariable("customerId") Long customerId){
         customerService.deleteCustomerById(customerId);
     }
     @PutMapping("{customerId}")
     public void updateCustomer(
-            @PathVariable("customerId") Integer customerId,
+            @PathVariable("customerId") Long customerId,
             @RequestBody CustomerUpdateRequest updateRequest
             ){
         customerService.updateCustomer(customerId, updateRequest);
