+++ src/main/java/com/springapp/customer/CustomerJDBCDataAccessService.java
@@ -29,7 +29,7 @@ public class CustomerJDBCDataAccessService implements CustomerDao{
     }
 
     @Override
-    public Optional<Customer> selectCustomerById(Integer id) {
+    public Optional<Customer> selectCustomerById(Long id) {
         var sql = """
                 SELECT id,name,email,age FROM customer
                 WHERE id = ?
@@ -65,7 +65,7 @@ public class CustomerJDBCDataAccessService implements CustomerDao{
     }
 
     @Override
-    public boolean existsPersonWithId(Integer id) {
+    public boolean existsPersonWithId(Long id) {
         var sql = """
                 SELECT count(*) FROM customer
                 WHERE id = ?
@@ -75,7 +75,7 @@ public class CustomerJDBCDataAccessService implements CustomerDao{
     }
 
     @Override
-    public void deleteCustomerById(Integer id) {
+    public void deleteCustomerById(Long id) {
         if(!existsPersonWithId(id)){
             throw new ResourceNotFoundException("customer with id [%s] not found".formatted(id));
         }
